+++ scanner/minimal_example.c
@@ -27,9 +27,9 @@ int main(int ac, const char* av[]) {
     memset(&file_info, '\0', sizeof(file_info));
     file = sf_open(av[i + 1], SFM_READ, &file_info);
 
-    sts[i] = ebur128_init((size_t) file_info.channels,
-                      (size_t) file_info.samplerate,
-                      EBUR128_MODE_I);
+    sts[i] = ebur128_init((unsigned) file_info.channels,
+                          (unsigned) file_info.samplerate,
+                          EBUR128_MODE_I);
 
     /* example: set channel map (note: see ebur128.h for the default map) */
     if (file_info.channels == 5) {
